Extracted creating test databases to testutils
